refactor: extract Layout interface from app.go
Separates rendering and navigation from state and message handling.
TabsLayout implements the Layout interface with all view/nav logic;
App retains thin wrapper methods so existing call sites and tests are
unchanged. Zero behaviour change.
