+++ internal/ui/layout.go
@@ -1,0 +1,12 @@
+package ui
+
+import tea "github.com/charmbracelet/bubbletea"
+
+// Layout arranges the app's screens and handles navigation for a specific UI paradigm.
+// All application state lives on App; Layout provides only method implementations.
+type Layout interface {
+	View(a App) string
+	HandleNav(msg tea.KeyMsg, a App) (App, tea.Cmd, bool)
+	DelegateUpdate(msg tea.Msg, a App) (App, tea.Cmd)
+	HasFocusedInput(a App) bool
+}
