+++ internal/ui/layout_tabs.go
@@ -1,0 +1,762 @@
+package ui
+
+import (
+	"fmt"
+	"strings"
+
+	tea "github.com/charmbracelet/bubbletea"
+	"github.com/charmbracelet/lipgloss"
+	"github.com/charmbracelet/x/ansi"
+	"github.com/ragnar/cyber-tui/internal/ui/imgview"
+	"github.com/ragnar/cyber-tui/internal/ui/theme"
+	"github.com/ragnar/cyber-tui/internal/ui/urlutil"
+	"github.com/ragnar/cyber-tui/internal/version"
+)
+
+var renderedVersionLine = theme.Subtle.Render("version " + version.Version + " (" + version.Commit + ")")
+
+// menuTabs is the ordered list of navigable screens.
+var menuTabs = []struct {
+	label string
+	s     screen
+}{
+	{"feed", screenFeed},
+	{"notifications", screenNotifications},
+	{"journal", screenJournal},
+	{"bookmarks", screenBookmarks},
+	{"guilds", screenGuilds},
+	{"topics", screenTopics},
+	{"profile", screenProfile},
+	{"settings", screenSettings},
+}
+
+// hint is a compact key+description pair shown in the status bar and help modal.
+type hint struct{ key, desc string }
+
+// sbStyle returns a bare style with the status-bar background.
+func sbStyle() lipgloss.Style {
+	return lipgloss.NewStyle().Background(theme.ColorDimGreen)
+}
+
+// renderHints formats a []hint slice as a compact styled string.
+func renderHints(hints []hint) string {
+	key := sbStyle().Foreground(theme.ColorCyan).Bold(true)
+	desc := sbStyle().Foreground(theme.ColorWhite)
+	sep := sbStyle().Foreground(theme.ColorMuted).Render(" · ")
+	parts := make([]string, 0, len(hints)*3)
+	for i, h := range hints {
+		if i > 0 {
+			parts = append(parts, sep)
+		}
+		parts = append(parts, key.Render(h.key))
+		if h.desc != "" {
+			parts = append(parts, desc.Render(" "+h.desc))
+		}
+	}
+	return strings.Join(parts, "")
+}
+
+// hintRows converts hints to modal row strings, skipping the "?" entry.
+// "↑↓" is expanded to "↑↓ / j/k" so the modal documents both navigation styles.
+func hintRows(hints []hint, rowFn func(string, string) string) []string {
+	rows := make([]string, 0, len(hints))
+	for _, h := range hints {
+		if h.key == "?" {
+			continue
+		}
+		key := h.key
+		if key == "↑↓" {
+			key = "↑↓ / j/k"
+		}
+		rows = append(rows, rowFn(key, h.desc))
+	}
+	return rows
+}
+
+// overlayCenter composites fg centered over bg using ANSI-aware string splicing.
+// Each line of fg replaces the corresponding characters in bg at the centered
+// position, preserving ANSI colour codes on both sides of the splice point.
+func overlayCenter(bg, fg string, bgW, bgH int) string {
+	fgW := lipgloss.Width(fg)
+	fgLines := strings.Split(fg, "\n")
+	fgH := len(fgLines)
+	bgLines := strings.Split(bg, "\n")
+
+	xOff := (bgW - fgW) / 2
+	yOff := (bgH - fgH) / 2
+	if xOff < 0 {
+		xOff = 0
+	}
+	if yOff < 0 {
+		yOff = 0
+	}
+
+	result := make([]string, len(bgLines))
+	copy(result, bgLines)
+
+	for i, fgLine := range fgLines {
+		bi := yOff + i
+		if bi < 0 || bi >= len(result) {
+			continue
+		}
+		bgLine := result[bi]
+		// Pad the background line if it's shorter than the splice end point.
+		bgLineW := ansi.StringWidth(bgLine)
+		needed := xOff + fgW
+		if bgLineW < needed {
+			bgLine += strings.Repeat(" ", needed-bgLineW)
+		}
+		left := ansi.Truncate(bgLine, xOff, "")
+		right := ansi.TruncateLeft(bgLine, xOff+fgW, "")
+		result[bi] = left + fgLine + right
+	}
+	return strings.Join(result, "\n")
+}
+
+// themeIndex returns the index of name in availableThemes, defaulting to 0.
+func themeIndex(name string) int {
+	for i, t := range availableThemes {
+		if t == name {
+			return i
+		}
+	}
+	return 0
+}
+
+// tabIndexOf returns the index of a.active within menuTabs, defaulting to 0.
+func tabIndexOf(a App) int {
+	for i, t := range menuTabs {
+		if t.s == a.active {
+			return i
+		}
+	}
+	return 0
+}
+
+// navigateTabBy computes the App state and load command for moving delta steps
+// through menuTabs from the current active screen.
+func navigateTabBy(a App, delta int) (App, tea.Cmd) {
+	if a.active == screenCMail {
+		a.cmail = a.cmail.CancelSubscription()
+	}
+	idx := (tabIndexOf(a) + delta + len(menuTabs)) % len(menuTabs)
+	a.active = menuTabs[idx].s
+	switch a.active {
+	case screenFeed:
+		a.feed = a.feed.SetFetching()
+		return a, a.loadFeedCmd()
+	case screenChatrooms:
+		return a, a.loadRoomsCmd()
+	case screenCMail:
+		return a, a.loadConvsCmd()
+	case screenProfile:
+		return a, a.loadProfileCmd()
+	case screenNotifications:
+		a.notifications = a.notifications.SetFetching()
+		return a, a.loadNotifsCmd()
+	case screenSettings:
+		return a, nil
+	case screenBookmarks:
+		if !a.bookmarks.IsLoaded() {
+			a.bookmarks = a.bookmarks.SetFetching()
+			return a, a.loadBookmarksCmd("")
+		}
+		return a, nil
+	case screenGuilds:
+		if !a.guilds.IsLoaded() {
+			a.guilds = a.guilds.SetFetching()
+			return a, a.loadGuildsCmd("")
+		}
+		return a, nil
+	case screenTopics:
+		if !a.topics.IsLoaded() {
+			a.topics = a.topics.SetFetching()
+			return a, a.loadTopicsCmd()
+		}
+		return a, nil
+	case screenJournal:
+		a.journal = a.journal.SetFetching()
+		return a, a.loadJournalCmd()
+	}
+	return a, nil
+}
+
+// TabsLayout implements the classic horizontal tab bar layout.
+type TabsLayout struct{}
+
+// View renders the full terminal output for the tabs layout.
+func (l TabsLayout) View(a App) string {
+	if a.active == screenLogin {
+		return a.login.View()
+	}
+	contentHeight := a.height - theme.ChromeHeight
+	content := lipgloss.NewStyle().Height(contentHeight).MaxHeight(contentHeight).Render(l.renderActiveScreen(a))
+	base := lipgloss.JoinVertical(lipgloss.Left,
+		l.renderTabBar(a),
+		"",
+		content,
+		l.renderBottomBar(a),
+	)
+	if a.themePickerOpen {
+		return overlayCenter(base, l.renderThemePicker(a), a.width, a.height)
+	}
+	if a.helpModalOpen {
+		return overlayCenter(base, l.renderHelpModal(a), a.width, a.height)
+	}
+	if a.urlPickerOpen {
+		return overlayCenter(base, l.renderURLPicker(a), a.width, a.height)
+	}
+	if a.imageModalOpen {
+		textModal := l.renderImageModal(a)
+		composed := overlayCenter(base, textModal, a.width, a.height)
+		// Compute the same offsets overlayCenter used so we can position
+		// the image sequence inside the border without embedding it in the
+		// overlay string (which would corrupt overlayCenter's ANSI splicing).
+		modalW := lipgloss.Width(textModal)
+		modalH := len(strings.Split(textModal, "\n"))
+		xOff := (a.width - modalW) / 2
+		yOff := (a.height - modalH) / 2
+		if xOff < 0 {
+			xOff = 0
+		}
+		if yOff < 0 {
+			yOff = 0
+		}
+		// theme.ActiveBorder: 1-char border + 1-char horizontal padding on each
+		// side. Image content therefore starts 2 cols right of the border edge.
+		// ANSI cursor sequences are 1-indexed; the border top row is yOff+1.
+		imgRow := yOff + 2
+		imgCol := xOff + 3
+		return composed + fmt.Sprintf("\x1b[%d;%dH%s\x1b[%d;1H", imgRow, imgCol, a.imageModalEncoded, a.height)
+	}
+	if a.imageNeedsCleanup && a.graphicsProtocol == imgview.ProtocolKitty {
+		// Inject the Kitty delete-all command onto the line that held the modal's
+		// top border so Bubble Tea's diff renderer delivers it to the terminal.
+		modalH := a.imageModalRows + 2
+		yOff := (a.height - modalH) / 2
+		if yOff < 0 {
+			yOff = 0
+		}
+		lines := strings.Split(base, "\n")
+		if yOff < len(lines) {
+			lines[yOff] = "\x1b_Ga=d,d=A\x1b\\" + lines[yOff]
+		}
+		return strings.Join(lines, "\n")
+	}
+	return base
+}
+
+// HandleNav processes navigation key presses for the tabs layout:
+// number shortcuts 1–8 and left/right arrow tab cycling.
+func (l TabsLayout) HandleNav(msg tea.KeyMsg, a App) (App, tea.Cmd, bool) {
+	switch msg.String() {
+	case "1":
+		if a.active != screenLogin {
+			a.cmail = a.cmail.CancelSubscription()
+			a.active = screenFeed
+			a.feed = a.feed.SetFetching()
+			return a, a.loadFeedCmd(), true
+		}
+	case "2":
+		if a.active != screenLogin {
+			a.cmail = a.cmail.CancelSubscription()
+			a.active = screenNotifications
+			a.notifications = a.notifications.SetFetching()
+			return a, a.loadNotifsCmd(), true
+		}
+	case "3":
+		if a.active != screenLogin {
+			a.cmail = a.cmail.CancelSubscription()
+			a.active = screenJournal
+			a.journal = a.journal.SetFetching()
+			return a, a.loadJournalCmd(), true
+		}
+	case "4":
+		if a.active != screenLogin {
+			a.cmail = a.cmail.CancelSubscription()
+			a.active = screenBookmarks
+			if !a.bookmarks.IsLoaded() {
+				a.bookmarks = a.bookmarks.SetFetching()
+				return a, a.loadBookmarksCmd(""), true
+			}
+			return a, nil, true
+		}
+	case "5":
+		if a.active != screenLogin {
+			a.cmail = a.cmail.CancelSubscription()
+			a.active = screenGuilds
+			if !a.guilds.IsLoaded() {
+				a.guilds = a.guilds.SetFetching()
+				return a, a.loadGuildsCmd(""), true
+			}
+			return a, nil, true
+		}
+	case "6":
+		if a.active != screenLogin {
+			a.cmail = a.cmail.CancelSubscription()
+			a.active = screenTopics
+			if !a.topics.IsLoaded() {
+				a.topics = a.topics.SetFetching()
+				return a, a.loadTopicsCmd(), true
+			}
+			return a, nil, true
+		}
+	case "7":
+		if a.active != screenLogin {
+			a.cmail = a.cmail.CancelSubscription()
+			a.active = screenProfile
+			return a, a.loadProfileCmd(), true
+		}
+	case "8":
+		if a.active != screenLogin {
+			a.cmail = a.cmail.CancelSubscription()
+			a.active = screenSettings
+			return a, nil, true
+		}
+	case "left":
+		if a.active != screenLogin && a.active != screenPostDetail && a.focus == focusMenu {
+			var cmd tea.Cmd
+			a, cmd = navigateTabBy(a, -1)
+			return a, cmd, true
+		}
+	case "right":
+		if a.active != screenLogin && a.active != screenPostDetail && a.focus == focusMenu {
+			var cmd tea.Cmd
+			a, cmd = navigateTabBy(a, +1)
+			return a, cmd, true
+		}
+	}
+	return a, nil, false
+}
+
+// DelegateUpdate routes a tea.Msg to the currently active screen model.
+func (l TabsLayout) DelegateUpdate(msg tea.Msg, a App) (App, tea.Cmd) {
+	var cmd tea.Cmd
+	switch a.active {
+	case screenLogin:
+		a.login, cmd = a.login.Update(msg)
+	case screenFeed:
+		a.feed, cmd = a.feed.Update(msg)
+	case screenChatrooms:
+		a.chatrooms, cmd = a.chatrooms.Update(msg)
+	case screenCMail:
+		a.cmail, cmd = a.cmail.Update(msg)
+	case screenProfile:
+		a.profile, cmd = a.profile.Update(msg)
+	case screenPostDetail:
+		a.postDetail, cmd = a.postDetail.Update(msg)
+	case screenNotifications:
+		a.notifications, cmd = a.notifications.Update(msg)
+	case screenSettings:
+		a.settingsScreen, cmd = a.settingsScreen.Update(msg)
+	case screenBookmarks:
+		a.bookmarks, cmd = a.bookmarks.Update(msg)
+	case screenGuilds:
+		a.guilds, cmd = a.guilds.Update(msg)
+	case screenTopics:
+		a.topics, cmd = a.topics.Update(msg)
+	case screenJournal:
+		a.journal, cmd = a.journal.Update(msg)
+	}
+	return a, cmd
+}
+
+// HasFocusedInput returns true when the active screen has a focused text input.
+func (l TabsLayout) HasFocusedInput(a App) bool {
+	switch a.active {
+	case screenChatrooms:
+		return a.chatrooms.InputFocused()
+	case screenCMail:
+		return a.cmail.InputFocused()
+	case screenPostDetail:
+		return a.postDetail.ComposeActive()
+	case screenFeed:
+		return a.feed.ComposeActive()
+	case screenGuilds:
+		return a.guilds.ComposeActive()
+	case screenProfile:
+		return a.profile.ComposeActive()
+	case screenJournal:
+		return a.journal.ComposeActive()
+	}
+	return false
+}
+
+func (l TabsLayout) renderTabBar(a App) string {
+	var tabs string
+	for _, t := range menuTabs {
+		label := t.label
+		if t.s == screenNotifications && a.polledUnreadCount > 0 {
+			label = fmt.Sprintf("%s (%d)", label, a.polledUnreadCount)
+		}
+		if a.active == t.s {
+			tabs += theme.ActiveTab.Render(label)
+		} else {
+			tabs += theme.Tab.Render(label)
+		}
+	}
+	logo := lipgloss.NewStyle().
+		Background(theme.ColorGreen).
+		Foreground(theme.ColorBackground).
+		Bold(true).
+		Padding(0, 1).
+		Render(a.logoText)
+	spacer := strings.Repeat(" ", max(0, a.width-lipgloss.Width(tabs)-lipgloss.Width(logo)))
+	return tabs + spacer + logo
+}
+
+func (l TabsLayout) renderActiveScreen(a App) string {
+	switch a.active {
+	case screenFeed:
+		return a.feed.View()
+	case screenChatrooms:
+		return a.chatrooms.View()
+	case screenCMail:
+		return a.cmail.View()
+	case screenProfile:
+		return a.profile.View()
+	case screenPostDetail:
+		return a.postDetail.View()
+	case screenNotifications:
+		return a.notifications.View()
+	case screenSettings:
+		return a.settingsScreen.View()
+	case screenBookmarks:
+		return a.bookmarks.View()
+	case screenGuilds:
+		return a.guilds.View()
+	case screenTopics:
+		return a.topics.View()
+	case screenJournal:
+		return a.journal.View()
+	}
+	return ""
+}
+
+func (l TabsLayout) renderBottomBar(a App) string {
+	if a.notifyText == "" {
+		return l.renderStatusBar(a)
+	}
+	return l.renderNotification(a)
+}
+
+func (l TabsLayout) renderNotification(a App) string {
+	color := theme.ColorGreen
+	prefix := "✓ "
+	switch a.notifyLevel {
+	case notifyWarn:
+		color = theme.ColorYellow
+		prefix = "! "
+	case notifyError:
+		color = theme.ColorRed
+		prefix = "✕ "
+	}
+	const suffix = "  (any key to dismiss)"
+	budget := a.width - lipgloss.Width(prefix) - lipgloss.Width(suffix) - 2
+	text := ansi.Truncate(a.notifyText, max(0, budget), "…")
+	return lipgloss.NewStyle().
+		Foreground(theme.ColorBackground).
+		Background(color).
+		Bold(true).
+		Width(a.width).
+		Padding(0, 1).
+		Render(prefix + text + suffix)
+}
+
+func (l TabsLayout) renderStatusBar(a App) string {
+	user := sbStyle().Foreground(theme.ColorCyan).Bold(true)
+	meta := sbStyle().Foreground(theme.ColorWhite)
+	sep := sbStyle().Foreground(theme.ColorMuted).Render(" · ")
+
+	densityLabel := "dense"
+	if a.relaxed {
+		densityLabel = "relaxed"
+	}
+	tzLabel := a.timezone
+	if tzLabel == "" {
+		tzLabel = "UTC"
+	}
+	timeFmt := a.settings.TimeDisplayFormat
+	if timeFmt == "" {
+		timeFmt = "datetime"
+	}
+
+	username := user.Render("@" + a.currentUser.Username)
+	infoItems := []string{
+		sep + meta.Render(densityLabel),
+		sep + meta.Render(theme.CurrentName()),
+		sep + meta.Render(tzLabel),
+		sep + meta.Render(timeFmt),
+	}
+
+	hints := l.screenHints(a)
+	const barPad = 2
+
+	measure := func(numInfo, numHints int) int {
+		left := lipgloss.Width(username)
+		for _, item := range infoItems[:numInfo] {
+			left += lipgloss.Width(item)
+		}
+		right := lipgloss.Width(renderHints(hints[:numHints]))
+		return left + right + barPad
+	}
+
+	numInfo := len(infoItems)
+	numHints := len(hints)
+	for numInfo >= 0 {
+		if measure(numInfo, numHints) <= a.width {
+			break
+		}
+		numInfo--
+	}
+	if numInfo < 0 {
+		numInfo = 0
+		for numHints > 1 && measure(0, numHints) > a.width {
+			numHints--
+		}
+	}
+
+	bg := sbStyle()
+	leftParts := []string{username}
+	leftParts = append(leftParts, infoItems[:numInfo]...)
+	left := lipgloss.JoinHorizontal(lipgloss.Top, leftParts...)
+	right := renderHints(hints[:numHints])
+	spacer := bg.Width(a.width - lipgloss.Width(left) - lipgloss.Width(right) - barPad).Render("")
+	bar := lipgloss.JoinHorizontal(lipgloss.Top, left, spacer, right)
+	return bg.Padding(0, 1).Render(bar)
+}
+
+func (l TabsLayout) screenHints(a App) []hint {
+	more := hint{"?", "more"}
+	switch a.active {
+	case screenFeed:
+		if a.feed.ComposeActive() {
+			return []hint{{"tab", "cycle"}, {"space", "toggle"}, {"Ctrl+s", "send"}, {"Esc", "cancel"}}
+		}
+		return []hint{{"↑↓", "navigate"}, {"enter", "open"}, {"r", "reply"}, {"n", "new"}, {"b", "bookmark"}, {"w", "watch"}, more}
+	case screenPostDetail:
+		if a.postDetail.ComposeActive() {
+			return []hint{{"Ctrl+s", "send"}, {"Esc", "cancel"}}
+		}
+		return []hint{{"↑↓", "navigate"}, {"r", "reply"}, {"b", "bookmark"}, {"w", "watch"}, {"esc", "back"}, more}
+	case screenProfile:
+		if a.profile.ComposeActive() {
+			return []hint{{"Ctrl+s", "save"}, {"Esc", "cancel"}, {"tab", "cycle"}}
+		}
+		if a.profile.IsReadOnly() {
+			return []hint{{"↑↓", "navigate"}, {"f", "follow"}, {"tab", "cycle"}, more}
+		}
+		return []hint{{"↑↓", "navigate"}, {"e", "edit"}, {"tab", "cycle"}, more}
+	case screenNotifications:
+		return []hint{{"↑↓", "navigate"}, {"enter", "open"}, {"m", "mark read"}, {"u", "toggle unread"}, more}
+	case screenJournal:
+		if a.journal.ComposeActive() {
+			return []hint{{"tab", "cycle"}, {"Ctrl+s", "save"}, {"Ctrl+p", "publish"}, {"Esc", "cancel"}}
+		}
+		return []hint{{"↑↓", "navigate"}, {"enter", "edit"}, {"n", "new"}, {"d", "delete"}, more}
+	case screenBookmarks:
+		return []hint{{"↑↓", "navigate"}, {"enter", "open"}, {"d", "delete"}, more}
+	case screenGuilds:
+		if a.guilds.ComposeActive() {
+			return []hint{{"tab", "cycle"}, {"Ctrl+s", "send"}, {"Esc", "cancel"}}
+		}
+		if a.guilds.IsBrowsingMembers() {
+			return []hint{{"↑↓", "navigate"}, {"enter", "view profile"}, {"esc", "back"}, more}
+		}
+		if a.guilds.IsBrowsingGuild() {
+			if a.guilds.IsConfirmingJoin() || a.guilds.IsConfirmingLeave() {
+				return []hint{{"y", "confirm"}, {"n/esc", "cancel"}}
+			}
+			hints := []hint{{"↑↓", "navigate"}, {"enter", "open"}, {"m", "members"}, {"n", "new thread"}, {"esc", "back"}}
+			d := a.guilds.GuildDetail()
+			if a.guilds.IsDetailLoaded() && !d.IsMember && a.currentUser.GuildSlug == "" {
+				hints = append(hints, hint{"J", "join"})
+			} else if a.guilds.IsDetailLoaded() && d.IsMember && d.Role != "founder" {
+				hints = append(hints, hint{"L", "leave"})
+			}
+			return append(hints, more)
+		}
+		return []hint{{"↑↓", "navigate"}, {"enter", "browse"}, more}
+	case screenTopics:
+		if a.topics.IsBrowsingTopic() {
+			return []hint{{"↑↓", "navigate"}, {"enter", "open"}, {"esc", "back"}, more}
+		}
+		return []hint{{"↑↓", "navigate"}, {"enter", "browse"}, {"esc", "back"}, more}
+	case screenSettings:
+		base := []hint{{"↑↓", "navigate"}, {"space", "toggle"}, {"tab", "cycle"}, more}
+		if a.settingsScreen.IsDirty() {
+			return append([]hint{{"Ctrl+s", "save"}, {"Esc", "revert"}}, base...)
+		}
+		return base
+	case screenCMail:
+		return []hint{{"← →", "switch pane"}, {"j/k", "navigate"}, {"enter", "send"}, more}
+	}
+	return []hint{more}
+}
+
+func (l TabsLayout) renderThemePicker(a App) string {
+	title := theme.Title.Render("theme")
+	var items []string
+	for i, name := range availableThemes {
+		if i == a.themePickerCursor {
+			items = append(items, theme.Highlight.Render("▸ "+name))
+		} else {
+			items = append(items, theme.Subtle.Render("  "+name))
+		}
+	}
+	hint := theme.Subtle.Render("↑↓ preview   enter save   esc cancel")
+	body := lipgloss.JoinVertical(lipgloss.Left,
+		title,
+		"",
+		lipgloss.JoinVertical(lipgloss.Left, items...),
+		"",
+		hint,
+	)
+	return theme.ActiveBorder.Render(body)
+}
+
+func (l TabsLayout) renderHelpModal(a App) string {
+	title := theme.Title.Render("shortcuts")
+	sectionStyle := theme.Subtle.Bold(true)
+	row := func(key, desc string) string {
+		k := theme.Highlight.Render(fmt.Sprintf("%-14s", key))
+		return lipgloss.JoinHorizontal(lipgloss.Top, k, theme.Subtle.Render(desc))
+	}
+
+	globalSection := lipgloss.JoinVertical(lipgloss.Left,
+		sectionStyle.Render("global"),
+		row("1-7", "feed · notifs · journal · bookmarks · topics · profile · settings"),
+		row("← →", "cycle tabs"),
+		row("t", "theme"),
+		row("v", "density"),
+		row("o", "open url"),
+		row("q", "quit"),
+	)
+
+	section := func(title string, extra ...string) string {
+		parts := append([]string{sectionStyle.Render(title)}, hintRows(l.screenHints(a), row)...)
+		parts = append(parts, extra...)
+		return lipgloss.JoinVertical(lipgloss.Left, parts...)
+	}
+
+	var localSection string
+	switch a.active {
+	case screenFeed:
+		if a.feed.ComposeActive() {
+			localSection = section("feed (compose)", row("Enter", "paragraph"))
+		} else {
+			localSection = section("feed",
+				row("p", "view profile"),
+				row("d", "delete own"),
+			)
+		}
+	case screenPostDetail:
+		if a.postDetail.ComposeActive() {
+			localSection = section("post detail (compose)", row("Enter", "paragraph"))
+		} else {
+			localSection = section("post detail",
+				row("d", "delete own"),
+				row("p", "view profile"),
+			)
+		}
+	case screenProfile:
+		if a.profile.ComposeActive() {
+			localSection = section("profile (editing)")
+		} else if a.profile.IsReadOnly() {
+			localSection = section("profile",
+				row("enter", "open"),
+			)
+		} else {
+			localSection = section("profile (own)",
+				row("enter", "open"),
+			)
+		}
+	case screenNotifications:
+		localSection = section("notifications",
+			row("M", "mark all read"),
+			row("p", "view profile"),
+		)
+	case screenJournal:
+		if a.journal.ComposeActive() {
+			localSection = section("journal (editing)", row("Enter", "paragraph"))
+		} else {
+			localSection = section("journal",
+				row("h", "revision history"),
+			)
+		}
+	case screenBookmarks:
+		localSection = section("bookmarks")
+	case screenGuilds:
+		if a.guilds.ComposeActive() {
+			localSection = section("guilds (compose)", row("Enter", "paragraph"))
+		} else if a.guilds.IsBrowsingMembers() {
+			localSection = section("guilds (members)", row("enter", "view profile"))
+		} else if a.guilds.IsBrowsingGuild() {
+			localSection = section("guilds (browsing)", row("n", "new thread"), row("m", "members"))
+		} else {
+			localSection = section("guilds")
+		}
+	case screenTopics:
+		if a.topics.IsBrowsingTopic() {
+			localSection = section("topics (browsing)")
+		} else {
+			localSection = section("topics")
+		}
+	case screenSettings:
+		t := "settings"
+		if a.settingsScreen.IsDirty() {
+			t = "settings (unsaved changes)"
+		}
+		localSection = section(t)
+	case screenCMail:
+		localSection = section("c-mail")
+	}
+
+	body := lipgloss.JoinVertical(lipgloss.Left,
+		title,
+		"",
+		globalSection,
+		"",
+		localSection,
+		"",
+		theme.Subtle.Render("any key · close"),
+		renderedVersionLine,
+	)
+	return theme.ActiveBorder.Render(body)
+}
+
+func (l TabsLayout) renderURLPicker(a App) string {
+	title := theme.Title.Render("open url")
+	items := make([]string, len(a.urlPickerItems))
+	for i, u := range a.urlPickerItems {
+		display := u
+		if urlutil.IsImageURL(u) &&
+			a.graphicsProtocol != imgview.ProtocolNone &&
+			a.imageViewer != "browser" {
+			display = "[img] " + display
+		}
+		if len(display) > 60 {
+			display = display[:57] + "..."
+		}
+		if i == a.urlPickerCursor {
+			items[i] = theme.Highlight.Render("▸ " + display)
+		} else {
+			items[i] = theme.Subtle.Render("  " + display)
+		}
+	}
+	hint := theme.Subtle.Render("↑↓ select   enter open   esc cancel")
+	rows := append([]string{title, ""}, items...)
+	rows = append(rows, "", hint)
+	return theme.ActiveBorder.Render(lipgloss.JoinVertical(lipgloss.Left, rows...))
+}
+
+// renderImageModal returns the bordered text-only shell for the image overlay.
+// The image escape sequence is injected separately in View() via ANSI cursor movement.
+func (l TabsLayout) renderImageModal(a App) string {
+	blankLine := strings.Repeat(" ", a.imageModalCols)
+	lines := make([]string, a.imageModalRows)
+	for i := range lines {
+		lines[i] = blankLine
+	}
+	return theme.ActiveBorder.Render(strings.Join(lines, "\n"))
+}
